Show warning instead of error when trying to remove nonexistent daily_plan_recipe or recipe_ingredient

+++ app/models/mixins/recipes/recipe_ingredients.py
@@ -31,4 +31,8 @@ class RecipeIngredientMixin:
 
     def remove_ingredient(self, ingredient):
         rhi = RecipeHasIngredient.load_by_recipe_and_ingredient(self, ingredient)
+        if not rhi:
+            return False
+
         rhi.delete()
+        return True
